rename permission

+++ src/controllers/backend/DefaultController.php
@@ -46,6 +46,9 @@ class DefaultController extends Controller
      */
     public function actionIndex()
     {
+        if (!\Yii::$app->user->can('menuBackendDefaultIndex')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $searchModel = new MenuSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -63,6 +66,9 @@ class DefaultController extends Controller
      */
     public function actionView($id)
     {
+        if (!\Yii::$app->user->can('menuBackendDefaultView')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -75,6 +81,9 @@ class DefaultController extends Controller
      */
     public function actionCreate()
     {
+        if (!\Yii::$app->user->can('menuBackendDefaultCreate')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = new Menu();
 
         if ($this->request->isPost) {
@@ -99,6 +108,9 @@ class DefaultController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!\Yii::$app->user->can('menuBackendDefaultUpdate')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -119,6 +131,9 @@ class DefaultController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!\Yii::$app->user->can('menuBackendDefaultDelete')) {
+            throw new \yii\web\ForbiddenHttpException(Module::t('You are not allowed to access this page.'));
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
